% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{graphic_heterozygosity}
\alias{graphic_heterozygosity}
\title{Graphic heterozygousity}
\usage{
graphic_heterozygosity(pop)
}
\arguments{
\item{pop}{Data from the function \code{simulation()}}
}
\value{
The described graphic
}
\description{
Creates a graphic that shows the heterozygosity of the number of alleles over
generations.
}
\examples{
p <- simulation()
h <- graphic_heterozygosity(pop = p)
}
