% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{graphic_individuals}
\alias{graphic_individuals}
\title{Graphic realizations}
\usage{
graphic_individuals(pop, limit = -1)
}
\arguments{
\item{pop}{Data from the function \code{simulation()}}

\item{limit}{The number of realizations that is included in the graphic as a maximum.
The mean is always built over all realizations. Default: -1, no limit.}
}
\value{
The described graphic
}
\description{
Creates a graphic that shows the evolution of the simulated populations. The mean of
the number of alleles is added.
}
\examples{
p <- simulation()
h <- graphic_individuals(pop = p)
}
