% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{heterozygous_individuals}
\alias{heterozygous_individuals}
\title{Heterozygosity}
\usage{
heterozygous_individuals(pop)
}
\arguments{
\item{pop}{The array containing the simulation from \code{simulation()}}
}
\value{
An array that indicates whether an individual is heterozygous
}
\description{
Creates an array, which contains a \code{1} for every heterozygous and a \code{0} for
every homozygous individual. The array has a similar structure like that after the
simulation, but the first dimension is only the number of individuals.
}
\examples{
p <- simulation()
h <- heterozygous_individuals(pop = p)
}
