% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{init}
\alias{init}
\title{Initialization}
\usage{
init(N, X0, realizations, gens)
}
\arguments{
\item{N}{Number of individuals in the population}

\item{X0}{Number of occurrences for that allele you are looking at}

\item{realizations}{Number of independent runs of the simulaion}

\item{gens}{Number of generations which should be simulated}
}
\description{
Sets some values, which are needed for all other functions. Always run this
function in the beginning before using any other function of the package!
}
\examples{
init(N = 50, X0 = 50, realizations = 50, gens = 400)
}
