% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{simulation}
\alias{simulation}
\title{Simulation}
\usage{
simulation()
}
\description{
Creates an array, which contains all necessary information about the genetic
development. This array is needed to create the graphics. It can get quite large.
}
\details{
All necessary parameters need to be set by using \code{init()}.
}
\examples{
p <- simulation()
}
